refactor(dnd): reusable dnd functions
