+++ src/components/panel/LeftPanelComponentItem.vue
@@ -1,113 +1,19 @@
 <script>
-import { mapGetters, mapActions } from 'vuex'
 import { ClassPrefix } from '../../const'
-import {
-  addEvent,
-  removeEvent,
-  SelectorAttrComponent,
-  dragElement
-} from '../../utils'
-
+import DnDMixin from '../../mixin/dnd'
 const mainClass = `${ ClassPrefix.LEFT_PANEL }-component-items`
 const itemClass = `${ ClassPrefix.MAIN }__grid-item`
-const itemDraggingClass = `${ itemClass }--dragging`
 
 export default {
   name: 'leftPanelComponentItem',
+  mixins: [DnDMixin],
   props: {
     items: {
       type: Array,
       default: () => []
     }
   },
-  data () {
-    return {
-      dragState: {
-        element: undefined,
-        x: 0,
-        y: 0
-      }
-    }
-  },
-  computed: {
-    ...mapGetters([
-      'iframeDocument',
-      'iframeWindow',
-      'canvasScroll',
-      'componentDragging'
-    ])
-  },
-  methods: {
-    ...mapActions([
-      'hideLeftPanels',
-      'enableDragComponent',
-      'disableDragComponent'
-    ]),
-    /**
-     * Move dragging component position
-     *
-     * @param {DOMObject} target
-     * @param {Boolean} isIframe
-     * @return void
-     */
-    moveComponent (target, isIframe = true) {
-      const iframeWindow = isIframe
-        ? this.iframeWindow
-        : undefined
 
-      dragElement(target, {
-        iframeWindow,
-        state: this.dragState,
-        canvasScrollTop: this.canvasScroll.top
-      })
-    },
-    dragStart (event) {
-      const { target, pageX, pageY } = event
-
-      this.dragState.element = target.cloneNode(true)
-      this.dragState.element.style.pointerEvents = 'none'
-      this.dragState.element.classList.add(itemDraggingClass)
-
-      document.body.appendChild(this.dragState.element)
-
-      this.dragState.x = pageX
-      this.dragState.y = pageY
-
-      // Start event dragging on document iframe
-      addEvent(document, 'mousemove', this.dragMove, false)
-      addEvent(document, 'mouseup', this.dragEnd, false)
-      addEvent(this.iframeDocument, 'mousemove', this.dragMove, false)
-      addEvent(this.iframeDocument, 'mouseup', this.dragEnd, false)
-
-      const componentId = this.dragState.element.getAttribute(SelectorAttrComponent)
-      this.enableDragComponent(componentId)
-      this.hideLeftPanels()
-    },
-    dragMove (event) {
-      if (!this.componentDragging) return false
-
-      const { pageX, pageY } = event
-
-      this.dragState.x = pageX
-      this.dragState.y = pageY
-
-      const isIframe = event.target.ownerDocument !== document
-      // Move element UI
-      this.moveComponent(this.dragState.element, isIframe)
-    },
-    dragEnd (event, moving) {
-      this.dragState.element.remove()
-      this.disableDragComponent()
-
-      // remove event dragging on document iframe
-      removeEvent(document, 'mousemove', this.dragMove, false)
-      removeEvent(document, 'mouseup', this.dragEnd, false)
-      removeEvent(this.iframeDocument, 'mousemove', this.dragMove, false)
-      removeEvent(this.iframeDocument, 'mouseup', this.dragEnd, false)
-
-      // do something when dropped
-    }
-  },
   render (h) {
     const itemEls = this.items.map(item => {
       const styles = {
@@ -119,11 +25,15 @@ export default {
         'no-image': !item.settings.icon
       }
 
+      const mousedown = (event) => {
+        this.dragStart(event, 'component')
+      }
+
       const key = `${ itemClass }--${ item._id }`
 
       return <div
         class={ itemClass }
-        onMousedown={ this.dragStart }
+        onMousedown={ mousedown }
         data-uno-component={ item.settings.id }
         key={ key }
         >
